Update everything up to "Testing"
This updates most of the write-up to make sense for the Rust elective. A lot of this was just replacing Java references with Rust references, most of which was pretty straightforward but still needed to happen. I also tried to link types and methods to the relevant Rust docs throughout, which will hopefully help.

I included suggested key data types and some methods.