Update the testing write-up
This gets us up to the `bats` section.

As part of this, we switched the file names from having type `String` to having type `OsString`. We also switched to using strings with `.as_bytes()` to create the byte arrays for test packets in our example, which should make it easier for the students to create test packets.

Thanks to @JustusFluegel for those suggestions (and many others).